Added code to test collisions against the player paddles.
Reworked the includes on a bunch of files because I was getting linker errors. I don't know if it is related to the order of the includes but I know it has something to do with Graphical.h. For example in DrawObject.h if I change the 3 includes of the Geometry classes for Graphical.h (which includes those elements) I get a bunch of linker errors. The current config works so I'm saving it like it is right now.