Added text support. Font and Text classes have been added. Created methods in GraphicEngine to create the base classes with the backing implementation.
Changed everywhere a std::string* was needed to a no pointer type, strings are all going to be passed as value.

Added the SetCenterPos in the DrawObject base class that I forgot.

Added reference check to NULL in the DSprite class.

Added Set methods in FSize and FPosition to allow setting two new values instead of deleting and setting the pointer.

Added default DistanceFromType as OriginToOrigin in the FRectangle.

Added includes in the CommsFramework to make sure they are compiled to check for errors.