Fixed the memory leaks in the FileReader and PackageFile classes.
Changed the FileReader to the concept that a GetFileContent returns the data of a file and does NOT keep a copy of it. So callers need to hang on to that content. Removed the member newcontent and contentSize altogether from the FileReader.

Changed the memory management of PackageFile. Fixed the leaks when saving a file.