Reverted the solution's project to Visual Studio 2013 because SFML is compiled for VS2013 and will not work when used with the new toolset. The solution can still be opened in VS2015.
Fixed a bug with the Package file when dealing with files with null terminators inside.

Fixed a bug with the FileReader that was not opening the file in Binary mode.

Added a Utilities file with a function to dump binary data to a file.

Finished the method SetupSprite to load resources from the frames defined in the XML file. No framelist for now.

Added some test assets.