Moved the IsLastFrame to the BaseSprite class to allow checking anywhere if the sprite is done animating.
Removed the IsPlaying check when drawing a new frame. To ensure that even when a sprite is not called Play() we can still use SetFrame and stuff.

Added a Reset() function that makes a SetFrame(0)

Forgot to add a CurrentFrameIndex in the SetFrame call.

Added a state machine to the coin object to try some functionalities.