Completed the coin click implementation to make it spin.
Converted the mouse position to float just to ease compatibility with other functions accepting Vector<float>.

Started replacing calls that were returning pointers to a small primitive object to returning a value instead.

Replaced the IsPlaying check in Draw of DSprite because that was causing the sprite to go to the next frame even when it was not playing. It has nothing to do with the drawing of it.

Added an overload of IsPointInside taking a Vector2<float> as parameter.

Fixed a bug in Rectangle where the Top member was not getting the right value. Also changed the check of the IsPointInside to add equals.

Keyboard and Mouse are getting references to the GraphicEngine, in part this was due to the Mouse needing a reference to the window to calculate the mouse position relative to the 0,0 of the window.