Sprite clicks refactor.
Before this commit, BaseSprites would manually check if they were clicked. This would have caused problem when several sprites were on top of each other. The expected behavior is for the top sprite to receive the click and decide if the event is sent to the sprites beneath.

The new implementation has the GameEngine check which sprites were clicked and sets a flag on the DrawObject class (DrawObject now has the members related to the clicks. No more need to be a BaseSprite to recieve clicks, this helps the Text class). Now when a sprite checks for the Clicked() method, it checks it his flag has been set by the engine. After that, it can check which button was pressed and react appropriately.

The Engine checks the clicks in the Pre_Update and removes the click flags in the Post_Update method.

Details :

- Added a few forward declares for good measure.

- Added an include of GameEngine in the .cpp of the DrawObject class, putting the include in the .h file seems to introduce linking errors.

- Added an overload of Mouse->Clicked() with no specified button just to check if the mouse has a clicked button, no matter which.

- GridTile check for click is now done on the CoveredSprite because that's the sprite that is on top. Due to the new click propagation the covered sprite does not recieve the event. This temporarily prevents the player from covering a tile once revealed.