Added the new dialog images to the class. Forgot to change their ident name previously.
Fixed a few bugs in the z-index reordering : First sprites are now also reordered whenever a new sprite is added. SFMLTexts now start not visible.

Reversed the iteration of the Sprites list in the click detection to fix the wrong logical assumption that the lowest sprites in the list would be the first shown. They are the first to be rendered thus they are the ones that gets underneath all the other sprites. By going reverse we not test the top sprites first.

Fixed the logic in the game flow.