Added dealer cards on screen.
Started a refactor of the click detection system for sprites. They are now given a pointer to a structure fo a click info that contains what keys were pressed and the location.  This fixes a bug where the detection of what key was pressed would be deffered to much later. So it was possible for the mouse LEFT to be depressed when the moment of evaluation of the click came.