Some memory management.
Added a delay after a game to see the results.

Added some commented results in the BaseStack class in order to check if the list is empty when popping an item.

Added classes to manage sockets servers and clients.

Added a Utils.h file to contain some general global funtions or variables.