Implemented several suggestions from a code analysis by visual studio.
A big problem was found in the GraphicEngine where the target window would dereference a NULL in cases where the SFMLDrawable class did not implement sf::Drawable and could not pass a dynamic_cast.

Implemented a numbered string copy in the PackageFile to ensure that only 4 bytes are copied in the header signature.

Added init to NULL for several pointers.

Fixed a == to = in the SocketClient.

Implemented array deletes at a couple of places. New syntax that I did not know about.

Changed some refferencing logic in the ChangeWorkingDir of the XDirectory class. Not sure about that change and I did not test it.