Memory stability upgrade.
The problem is that the GameEngine and the GraphicEngine both have a list of objects that needs to be synchronized. In this case the objects were being deleted in one list but not the other. So I added an overload in the Engine to allow removing the objects from the GameEngine and the underlying GraphicEngine.

Code cleanup. Coded a manual delete of all objects of the DeckCards to bypass the ESP problem with templates.

Uncommented the BaseStack HasNext implementation. I'm not using it anywhere so I'm not sure it will generate ok.