Seems I fixed the click problem with a strike of the axe.
I added a method in BaseMouse that takes the current mouse state and applies it to the previous state. I took down the current implementation of UpdateMouseState. The function is still there but it's doing nothing.

The real work is done in the HandleEvent function of SDLMouse, this sets the mouse state to clicked or released. The BaseMouse will do the work of saving the past state.

The DoEventLoop has been moved to the top of the Play() function, it should have been there since the beginning since it is the code that sets the state used throughout the loop.