Added a new class that enables adding key/value pairs and save that to a file. Currently the implementation does not save but can add values by type.
Another class is GenType, short for generic type. This class can store a value as a void* and return it as it's correct type by storing an enum that corresponds to the value of the data. This is a kind of fake generic class for holding basic values used in the FileSave class.