Added a function to add a spritesheet into the texture repository. This will extract the sprites and add each of them into the repository.
Added an overload of CreateSprite to allow creating a sprite without an identifier.

Added a member TextureName to BaseTexture, this will allow finding a texture from the repository not by path but by name.

Changed the way to create a Spritesheet. The configFile path needs to be given at construction of the spritesheet.

Added files to the FantasyGame project for the map classes.