Added overloads in the IOUtilities file to handle checking path and creating a series of directories based on a path string.
Added some asserts in the IOUtilities code file.

Moved the functionality of the GetFilePathComponents to an overload to handle getting the components of a path from a string rather than a file.

Fixed a bug in the StringFunctions file where a buffer needed +1 to its size to account for the null terminator. Also moved the call to strtok AFTER the add in the list, otherwise the first element would be skipped.