Completed the implementation of the Packager and UnPackager.
Added a mechanism in the PackageFile, when adding directories of files, to add files with names relative to the root of the directory being added. For example when adding C:\temp\test and all the files inside. The file C:\temp\test\blabla\image.png will be named 'test\blabla\image.png ' if the package was created at the 'test' folder.

This has changed the structure of adding the file entries, they are now objects that tracks the directory that they were added with. Files being added store an empty directory so no changes there.

Added a StringSubtract method to subtract a string from another string. Very helpful when transforming a file path in a path relative to a parent directory.