Added a new class to handle encrypted package files. Started adding implementation for the overridden methods.
Added a method in PackageFile to group the AddFile functionality into a core function that would be called in the AddDirectory function so that all needs to add a file to the package file goes through that function.

Also decorated methods in PackageFile with virtual and added the members in the protected space rather than private.