Refactored the XmlReader class
Two sets of function now exists in the XmlReader class. The old one was renamed from GetNode to FindNode. "Find" is used to qualify that the class will search throughout the document for the node. Even if it has to step inside children and siblings.

The new function set is using the old name of GetNode. This is used to show that the query will only look through the children of the current node on which the function is called. Those functions will not go through sub-children so a chain of calls must be done for precise search.

- That change was needed because the old ways did not support having multiple nodes in the document with the same names. The methods would start at the beginning of the document and search through it and return the first result it found.