Sent the IsTimeForFrame function to public.
This was done because the GameEngine was running as fast as it could and we really only need to make an update at each 60 frame.

With that, the implementing system can sync the engine update on each frame.

Replaced the implementation of the IsTimeForFrame because we need to query the current frame time.