Moved the window and D2D object creation to inside the thread function.
Moving the WindowCreate function fixes the bug where the GetMessage never returned.

Switched the checked array in the GetNextFreeHwndIndex to use an array that was filled immediately by the current thread. This prevents problems where opening multiple windows in succession would return the same index because the threads were not started yet and did not add their window into the array.