Code changes for mandate #001. Texture names.
Added code in BaseSprite and SDLSprite to check if a sprite was given a name but no texture, allowing the engine to load the texture at a later time. Code is not finished yet. Still need to write the code to load and unload textures from sprites.

Test App code changes.