Work on the map init.
Created a InitializeDefaultMap function to setup a random grid of tiles. A new system of tile initialization will be implemented. Tiles won't create their own sprite because I don't want them to have the responsibility of setting their sprite size and messing with texture stuff. Instead they will set themselves a Coordinate and a TileIdentifier that will allow the game engine to setup the tile externally.