Added a primitives drawing subsystem.
This allows to draw basic rectangles and lines. Next will be more complex shapes like circles and ovals.

The primitives are stores in a separate list and drawn AFTER the other drawables. So the Z index of the primitives is ignored for now. Ideally they would share the z-index space with the sprites.

Added a .cpp file to the BaseGraphicEngine. I was opposed to this at first but this allows the primitive drawing engine to be almost completely implemented in the base class. Other functions that can fit in the BaseGraphicEngine might get migrated there in the future.