Rework of the texture creation pipeline. Document will follow.
In short, new functions in BaseGraphicEngine have been created to allow creating textures. The CreateTextureInstance has been sent to the protected section. That way, all calls to CreateTexture should include a texture name and/or a path. An overload with no name or path has been created to allow creating anonymous textures that can be loaded later. They cannot be traced by name so better hang to your BaseTexture instances if you want it back.

If each texture have a name, they can be defer loaded easily and deduplicated.