Emscripten compilation
Implemented a working workflow for H5 compilation with emscripten.

The src directory contains a script (emscripten_cmake.sh) to cmake a new build directory from which we can call make.

CMakeLists have been updates to avoid going into the UNIX branch. This was causing the find_library calls to fail and we don't need them anyway because the emscripten compiler takes care of those libraries anyway. The EMSCRIPTEN branch still looks for SDL libs but I think there is a way to omit them.

Added emscripten calls to GetTicks and an emscripten call to emscripten_set_main_loop to start the request_frame.

I can't get the cmake build to output an html file with the JS linked to it. So I have a dummy index.html that I copy to the output directory that needs to be linked to the correct Javascript of the target game. Scroll to the bottom and change the included js name.